Use uint16_t for slot index

+++ src/sacn/private/dmx_merger.h
@@ -38,7 +38,7 @@ extern "C" {
 
 typedef struct WinnerLookupKeys
 {
-  unsigned int slot_index;
+  uint16_t slot_index;
   source_id_t owner;
   uint8_t level;
   uint8_t priority;
@@ -86,21 +86,21 @@ bool merger_handle_in_use(int handle_val, void* cookie);
 bool source_handle_in_use(int handle_val, void* cookie);
 
 etcpal_error_t update_levels(MergerState* merger, SourceState* source, const uint8_t* new_values,
-                             size_t new_values_count);
-etcpal_error_t update_level(MergerState* merger, SourceState* source, unsigned int level_index, uint8_t level);
+                             uint16_t new_values_count);
+etcpal_error_t update_level(MergerState* merger, SourceState* source, uint16_t level_index, uint8_t level);
 etcpal_error_t update_per_address_priorities(MergerState* merger, SourceState* source,
-                                             const uint8_t* address_priorities, size_t address_priorities_count);
-etcpal_error_t update_priority(MergerState* merger, SourceState* source, unsigned int priority_index, uint8_t priority);
+                                             const uint8_t* address_priorities, uint16_t address_priorities_count);
+etcpal_error_t update_priority(MergerState* merger, SourceState* source, uint16_t priority_index, uint8_t priority);
 etcpal_error_t update_universe_priority(MergerState* merger, SourceState* source, uint8_t priority);
-WinnerLookupKeys get_winner_lookup_source_lower_bound_keys(unsigned int slot_index, SourceState* source);
-WinnerLookupKeys get_winner_lookup_source_upper_bound_keys(unsigned int slot_index, SourceState* source);
-WinnerLookupKeys get_winner_lookup_slot_lower_bound_keys(unsigned int slot_index);
-WinnerLookupKeys get_winner_lookup_slot_upper_bound_keys(unsigned int slot_index);
+WinnerLookupKeys get_winner_lookup_source_lower_bound_keys(uint16_t slot_index, SourceState* source);
+WinnerLookupKeys get_winner_lookup_source_upper_bound_keys(uint16_t slot_index, SourceState* source);
+WinnerLookupKeys get_winner_lookup_slot_lower_bound_keys(uint16_t slot_index);
+WinnerLookupKeys get_winner_lookup_slot_upper_bound_keys(uint16_t slot_index);
 etcpal_error_t update_winner_lookup(MergerState* merger, WinnerLookupKeys* current_keys_to_free,
                                     const WinnerLookupKeys* new_keys_to_copy);
-WinnerLookupKeys* get_current_keys(MergerState* merger, SourceState* source, unsigned int slot_index);
+WinnerLookupKeys* get_current_keys(MergerState* merger, SourceState* source, uint16_t slot_index);
 bool keys_valid(const WinnerLookupKeys* keys);
-void update_merge(MergerState* merger, unsigned int slot_index);
+void update_merge(MergerState* merger, uint16_t slot_index);
 void deinit_merger_state(MergerState* state);
 
 #ifdef __cplusplus
